SDK-537: Separate attribute and anchor tests
